+++ index.html
@@ -94,25 +94,25 @@
   <!-- ==== Scripts ==== -->
   
   <!-- ==== DATA & VIDEO ==== -->
-  <script src="js/videoPlayer.js"></script>
-  <script type="module" src="js/dataLinker.js"></script>
+  <script src="js/video/videoPlayer.js"></script>
+  <script type="module" src="js/data/dataLinker.js"></script>
 
   <!-- ==== FILTRES ==== -->
-  <!-- <script type="module" src="js/filtreOpen.js"></script> -->
+  <!-- <script type="module" src="js/ui/filtreOpen.js"></script> -->
 
   <!-- ==== VISUALISER ==== -->
-  <!-- <script type="module" src="js/textManagerLyricsAnimation.js"></script>  -->
-  <script type="module" src="js/textManagerLyricsAnimation02.js"></script> 
-  <!-- <script type="module" src="js/eventManager.js"></script>  -->
-  <script src="js/SHADER/videoShader.js"></script>
-  <!-- <script type="module" src="js/textManager.js"></script> 
-  <script type="module" src="js/textAnimations.js"></script>  -->
-  <!-- <script type="module" src="js/timeline.js"></script>  -->
+  <!-- <script type="module" src="js/ui/textManagerLyricsAnimation.js"></script>  -->
+  <script type="module" src="js/ui/lyricsDisplay.js"></script>
+  <!-- <script type="module" src="js/ui/eventManager.js"></script>  -->
+  <script src="js/video/shader/videoShader.js"></script>
+  <!-- <script type="module" src="js/ui/textManager.js"></script>
+  <script type="module" src="js/ui/textAnimations.js"></script>  -->
+  <!-- <script type="module" src="js/ui/timeline.js"></script>  -->
 
   <!-- ==== TIMELINE ==== -->
-  <!-- <script type="module" src="js/timeline02.js"></script>  -->
-  <script type="module" src="js/timeline03.js"></script> 
-  <!-- <script type="module" src="js/AudioReactiveType/audioReactiveType_02.js"></script>  -->
+  <!-- <script type="module" src="js/ui/timeline02.js"></script>  -->
+  <script type="module" src="js/ui/speakerTimeline.js"></script>
+  <!-- <script type="module" src="js/audio/reactiveType/audioReactiveType_02.js"></script>  -->
