+++ js/audio/reactiveType/audioReactiveType_02.js
@@ -1,4 +1,4 @@
-import { loadActiveArchiveData } from "../dataLinker.js";
+import { loadActiveArchiveData } from "../../data/dataLinker.js";
 
 document.addEventListener("DOMContentLoaded", async () => {
   const REM_IN_PX = 16;
