+++ js/ui/globalIndexViewer.js
