+++ js/ui/speakerTimeline.js
@@ -1,4 +1,4 @@
-import { loadActiveArchiveData } from "./dataLinker.js";
+import { loadActiveArchiveData } from "../data/dataLinker.js";
 
 loadActiveArchiveData()
   .then(data => {
@@ -76,10 +76,7 @@ loadActiveArchiveData()
     }
 
     function formatTime(sec) {
-      if (!isFinite(sec)) return "0:00";
-      const m = Math.floor(sec / 60);
-      const s = Math.floor(sec % 60);
-      return `${m}:${s < 10 ? "0" : ""}${s}`;
+      return `${Math.floor(sec)}`;
     }
 
     function updateTimelineCursor(currentTime) {
